+++ .vscode/settings.json
@@ -33,7 +33,47 @@
   "npm.packageManager": "yarn",
   // Extension: cSpell
   "cSpell.words": [
-    "Customizer"
+    "activite",
+    "Activité",
+    "Adresse",
+    "Ajouter",
+    "anonyme",
+    "bancaire",
+    "Banque",
+    "canadienne",
+    "céréales",
+    "collectif",
+    "commandite",
+    "création",
+    "Customizer",
+    "d'agrumes",
+    "Datde",
+    "datepicker",
+    "dénomination",
+    "détails",
+    "Entrer",
+    "Exonération",
+    "fourrages",
+    "Française",
+    "INCONNUE",
+    "industrielles",
+    "Informations",
+    "Juridique",
+    "légumineuses",
+    "limitée",
+    "nationalite",
+    "Nationalité",
+    "Poulina",
+    "Registre",
+    "responsabilité",
+    "Société",
+    "Societes",
+    "sociétés",
+    "tabler",
+    "Télephone",
+    "Tunisienne",
+    "unipersonnelle",
+    "Ville"
   ],
   "search-node-modules.path": "node_modules",
   "markdown.extension.toc.updateOnSave": false,
