+++ package.json
@@ -15,9 +15,9 @@
     "@casl/ability": "6.5.0",
     "@casl/react": "3.1.0",
     "@emotion/cache": "11.10.8",
-    "@emotion/react": "11.10.8",
+    "@emotion/react": "^11.11.3",
     "@emotion/server": "11.10.0",
-    "@emotion/styled": "11.10.8",
+    "@emotion/styled": "^11.11.0",
     "@fullcalendar/bootstrap5": "6.1.6",
     "@fullcalendar/common": "5.11.4",
     "@fullcalendar/core": "6.1.6",
@@ -29,7 +29,7 @@
     "@hookform/resolvers": "3.1.0",
     "@iconify/react": "4.1.0",
     "@mui/lab": "5.0.0-alpha.128",
-    "@mui/material": "5.12.2",
+    "@mui/material": "^5.15.10",
     "@mui/system": "5.12.1",
     "@mui/x-data-grid": "6.0.3",
     "@popperjs/core": "2.11.7",
