--- src/SharedComponents/FormIndvidual.tsx
@@ -1,436 +1,0 @@
-// ** React Imports
-import { ChangeEvent, Fragment, useState } from 'react'
-
-// ** MUI Imports
-import Box from '@mui/material/Box'
-import Card from '@mui/material/Card'
-import Grid from '@mui/material/Grid'
-import Avatar from '@mui/material/Avatar'
-import Button from '@mui/material/Button'
-import Divider from '@mui/material/Divider'
-import Stepper from '@mui/material/Stepper'
-import MenuItem from '@mui/material/MenuItem'
-import { styled } from '@mui/material/styles'
-import StepLabel from '@mui/material/StepLabel'
-import TextField from '@mui/material/TextField'
-import Typography from '@mui/material/Typography'
-import IconButton from '@mui/material/IconButton'
-import InputLabel from '@mui/material/InputLabel'
-import CardContent from '@mui/material/CardContent'
-import FormControl from '@mui/material/FormControl'
-import OutlinedInput from '@mui/material/OutlinedInput'
-import MuiStep, { StepProps } from '@mui/material/Step'
-import InputAdornment from '@mui/material/InputAdornment'
-import Select, { SelectChangeEvent } from '@mui/material/Select'
-
-// ** Third Party Imports
-import toast from 'react-hot-toast'
-
-// ** Icon Imports
-import Icon from 'src/@core/components/icon'
-
-// ** Custom Components Imports
-import StepperCustomDot from './StepperCustomDot'
-import CustomAvatar from 'src/@core/components/mui/avatar'
-
-// ** Util Import
-import { hexToRGBA } from 'src/@core/utils/hex-to-rgba'
-
-// ** Styled Component
-import StepperWrapper from 'src/@core/styles/mui/stepper'
-
-interface State {
-  password: string
-  password2: string
-  showPassword: boolean
-  showPassword2: boolean
-}
-
-const steps = [
-  {
-    icon: 'tabler:home',
-    title: 'les détails de société',
-    subtitle: 'Entrer les détails'
-  },
-  {
-    icon: 'tabler:user',
-    title: 'Personal Info',
-    subtitle: 'Setup Information'
-  },
-  {
-    icon: 'tabler:link',
-    title: 'Social Links',
-    subtitle: 'Add Social Links'
-  }
-]
-
-const Step = styled(MuiStep)<StepProps>(({ theme }) => ({
-  paddingLeft: theme.spacing(4),
-  paddingRight: theme.spacing(4),
-  '&:first-of-type': {
-    paddingLeft: 0
-  },
-  '&:last-of-type': {
-    paddingRight: 0
-  },
-  '& .MuiStepLabel-iconContainer': {
-    display: 'none'
-  },
-  '& .step-subtitle': {
-    color: `${theme.palette.text.disabled} !important`
-  },
-  '& + svg': {
-    color: theme.palette.text.disabled
-  },
-  '&.Mui-completed .step-title': {
-    color: theme.palette.text.disabled
-  },
-  '&.Mui-completed + svg': {
-    color: theme.palette.primary.main
-  }
-}))
-
-const StepperCustomHorizontal = () => {
-  // ** States
-  //const [email, setEmail] = useState<string>('')
-  const [dénomination, setDénomination] = useState<string>('')
-  const [juridique, setJuridique] = useState<string>('')
-  const [nom_commercial, setNomCommercial] = useState<string>('')
-  const [google, setGoogle] = useState<string>('')
-  const [country, setCountry] = useState<string>('')
-  const [twitter, setTwitter] = useState<string>('')
-  const [lastName, setLastName] = useState<string>('')
-  const [facebook, setFacebook] = useState<string>('')
-  const [linkedIn, setLinkedIn] = useState<string>('')
-  const [firstName, setFirstName] = useState<string>('')
-  const [activeStep, setActiveStep] = useState<number>(0)
-  const [language, setLanguage] = useState<string[]>([])
-  const [state, setState] = useState<State>({
-    password: '',
-    password2: '',
-    showPassword: false,
-    showPassword2: false
-  })
-
-  // Handle Stepper
-  const handleBack = () => {
-    setActiveStep(prevActiveStep => prevActiveStep - 1)
-  }
-  const handleNext = () => {
-    setActiveStep(prevActiveStep => prevActiveStep + 1)
-    if (activeStep === steps.length - 1) {
-      toast.success('Form Submitted')
-    }
-  }
-  const handleReset = () => {
-    setJuridique('')
-    setDénomination('')
-
-    //setEmail('')
-    setGoogle('')
-    setCountry('')
-    setTwitter('')
-    setLastName('')
-    setFacebook('')
-    setLinkedIn('')
-    setLanguage([])
-    setFirstName('')
-    setActiveStep(0)
-    setState({ ...state, password: '', password2: '' })
-  }
-
-  // Handle Password
-  const handlePasswordChange = (prop: keyof State) => (event: ChangeEvent<HTMLInputElement>) => {
-    setState({ ...state, [prop]: event.target.value })
-  }
-  const handleClickShowPassword = () => {
-    setState({ ...state, showPassword: !state.showPassword })
-  }
-
-  // Handle Confirm Password
-  const handleConfirmChange = (prop: keyof State) => (event: ChangeEvent<HTMLInputElement>) => {
-    setState({ ...state, [prop]: event.target.value })
-  }
-  const handleClickShowConfirmPassword = () => {
-    setState({ ...state, showPassword2: !state.showPassword2 })
-  }
-
-  // Handle Language
-  const handleSelectChange = (event: SelectChangeEvent<string[]>) => {
-    setLanguage(event.target.value as string[])
-  }
-
-  const getStepContent = (step: number) => {
-    switch (step) {
-      case 0:
-        return (
-          <Fragment>
-            <Grid item xs={12} sm={6}>
-              <p>Forme juridique :</p>
-              <TextField
-                fullWidth
-                value={juridique}
-                placeholder='Société en participation'
-                onChange={e => setJuridique(e.target.value)}
-              />
-            </Grid>
-              <Grid item xs={12} sm={6}>
-              <p>Dénomination :</p>
-              <TextField
-                fullWidth
-                onChange={e => setDénomination(e.target.value)}
-              />
-            </Grid>
-            <Grid item xs={12} sm={6}>
-              <p>Nom commercial :</p>
-              <TextField
-                fullWidth
-                onChange={e => setNomCommercial(e.target.value)}
-              />
-            </Grid>
-
-            <Grid item xs={12} sm={6}>
-              <FormControl fullWidth>
-                <InputLabel htmlFor='stepper-custom-horizontal-account-password'>Password</InputLabel>
-                <OutlinedInput
-                  label='Password'
-                  value={state.password}
-                  id='stepper-custom-horizontal-account-password'
-                  onChange={handlePasswordChange('password')}
-                  type={state.showPassword ? 'text' : 'password'}
-                  endAdornment={
-                    <InputAdornment position='end'>
-                      <IconButton
-                        edge='end'
-                        onClick={handleClickShowPassword}
-                        onMouseDown={e => e.preventDefault()}
-                        aria-label='toggle password visibility'
-                      >
-                        <Icon icon={state.showPassword ? 'tabler:eye' : 'tabler:eye-off'} />
-                      </IconButton>
-                    </InputAdornment>
-                  }
-                />
-              </FormControl>
-            </Grid>
-            <Grid item xs={12} sm={6}>
-              <FormControl fullWidth>
-                <InputLabel htmlFor='stepper-custom-horizontal-account-password-2'>Confirm Password</InputLabel>
-                <OutlinedInput
-                  value={state.password2}
-                  label='Confirm Password'
-                  id='stepper-custom-horizontal-account-password-2'
-                  onChange={handleConfirmChange('password2')}
-                  type={state.showPassword2 ? 'text' : 'password'}
-                  endAdornment={
-                    <InputAdornment position='end'>
-                      <IconButton
-                        edge='end'
-                        onMouseDown={e => e.preventDefault()}
-                        aria-label='toggle password visibility'
-                        onClick={handleClickShowConfirmPassword}
-                      >
-                        <Icon icon={state.showPassword2 ? 'tabler:eye' : 'tabler:eye-off'} />
-                      </IconButton>
-                    </InputAdornment>
-                  }
-                />
-              </FormControl>
-            </Grid>
-          </Fragment>
-        )
-      case 1:
-        return (
-          <Fragment key={step}>
-            <Grid item xs={12} sm={6}>
-              <TextField
-                fullWidth
-                label='First Name'
-                placeholder='Leonard'
-                value={firstName}
-                onChange={e => setFirstName(e.target.value)}
-              />
-            </Grid>
-            <Grid item xs={12} sm={6}>
-              <TextField
-                fullWidth
-                label='Last Name'
-                placeholder='Carter'
-                value={lastName}
-                onChange={e => setLastName(e.target.value)}
-              />
-            </Grid>
-            <Grid item xs={12} sm={6}>
-              <FormControl fullWidth>
-                <InputLabel id='stepper-custom-horizontal-personal-select-label'>Country</InputLabel>
-                <Select
-                  label='Country'
-                  value={country}
-                  id='stepper-custom-horizontal-personal-select'
-                  onChange={e => setCountry(e.target.value)}
-                  labelId='stepper-custom-horizontal-personal-select-label'
-                >
-                  <MenuItem value='UK'>UK</MenuItem>
-                  <MenuItem value='USA'>USA</MenuItem>
-                  <MenuItem value='Australia'>Australia</MenuItem>
-                  <MenuItem value='Germany'>Germany</MenuItem>
-                </Select>
-              </FormControl>
-            </Grid>
-            <Grid item xs={12} sm={6}>
-              <FormControl fullWidth>
-                <InputLabel id='stepper-custom-horizontal-personal-multiple-select-label'>Language</InputLabel>
-                <Select
-                  multiple
-                  value={language}
-                  onChange={handleSelectChange}
-                  id='stepper-custom-horizontal-personal-multiple-select'
-                  labelId='stepper-custom-horizontal-personal-multiple-select-label'
-                  input={<OutlinedInput label='Language' id='stepper-custom-horizontal-select-multiple-language' />}
-                >
-                  <MenuItem value='English'>English</MenuItem>
-                  <MenuItem value='French'>French</MenuItem>
-                  <MenuItem value='Spanish'>Spanish</MenuItem>
-                  <MenuItem value='Portuguese'>Portuguese</MenuItem>
-                  <MenuItem value='Italian'>Italian</MenuItem>
-                  <MenuItem value='German'>German</MenuItem>
-                  <MenuItem value='Arabic'>Arabic</MenuItem>
-                </Select>
-              </FormControl>
-            </Grid>
-          </Fragment>
-        )
-      case 2:
-        return (
-          <Fragment key={step}>
-            <Grid item xs={12} sm={6}>
-              <TextField
-                fullWidth
-                label='Twitter'
-                value={twitter}
-                onChange={e => setTwitter(e.target.value)}
-                placeholder='https://twitter.com/carterLeonard'
-              />
-            </Grid>
-            <Grid item xs={12} sm={6}>
-              <TextField
-                fullWidth
-                label='Facebook'
-                value={facebook}
-                onChange={e => setFacebook(e.target.value)}
-                placeholder='https://facebook.com/carterLeonard'
-              />
-            </Grid>
-            <Grid item xs={12} sm={6}>
-              <TextField
-                fullWidth
-                label='Google+'
-                value={google}
-                onChange={e => setGoogle(e.target.value)}
-                placeholder='https://plus.google.com/carterLeonard'
-              />
-            </Grid>
-            <Grid item xs={12} sm={6}>
-              <TextField
-                fullWidth
-                label='LinkedIn'
-                value={linkedIn}
-                onChange={e => setLinkedIn(e.target.value)}
-                placeholder='https://linkedin.com/carterLeonard'
-              />
-            </Grid>
-          </Fragment>
-        )
-      default:
-        return 'Unknown Step'
-    }
-  }
-
-  const renderContent = () => {
-    if (activeStep === steps.length) {
-      return (
-        <>
-          <Typography>All steps are completed!</Typography>
-          <Box sx={{ mt: 4, display: 'flex', justifyContent: 'flex-end' }}>
-            <Button size='large' variant='contained' onClick={handleReset}>
-              Reset
-            </Button>
-          </Box>
-        </>
-      )
-    } else {
-      return (
-        <form onSubmit={e => e.preventDefault()}>
-          <Grid container spacing={5}>
-            <Grid item xs={12}>
-              <Typography variant='body2' sx={{ fontWeight: 600, color: 'text.primary' }}>
-                {steps[activeStep].title}
-              </Typography>
-              <Typography variant='caption' component='p'>
-                {steps[activeStep].subtitle}
-              </Typography>
-            </Grid>
-            {getStepContent(activeStep)}
-            <Grid item xs={12} sx={{ display: 'flex', justifyContent: 'space-between' }}>
-              <Button
-                size='large'
-                variant='outlined'
-                color='secondary'
-                disabled={activeStep === 0}
-                onClick={handleBack}
-              >
-                Back
-              </Button>
-              <Button size='large' variant='contained' onClick={handleNext}>
-                {activeStep === steps.length - 1 ? 'Submit' : 'Next'}
-              </Button>
-            </Grid>
-          </Grid>
-        </form>
-      )
-    }
-  }
-
-  return (
-    <Card>
-      <CardContent>
-        <StepperWrapper>
-          <Stepper activeStep={activeStep} connector={<Icon icon='tabler:chevron-right' />}>
-            {steps.map((step, index) => {
-              const RenderAvatar = activeStep >= index ? CustomAvatar : Avatar
-
-              return (
-                <Step key={index}>
-                  <StepLabel StepIconComponent={StepperCustomDot}>
-                    <div className='step-label'>
-                      <RenderAvatar
-                        variant='rounded'
-                        {...(activeStep >= index && { skin: 'light' })}
-                        {...(activeStep === index && { skin: 'filled' })}
-                        {...(activeStep >= index && { color: 'primary' })}
-                        sx={{
-                          ...(activeStep === index && { boxShadow: theme => theme.shadows[3] }),
-                          ...(activeStep > index && { color: theme => hexToRGBA(theme.palette.primary.main, 0.4) })
-                        }}
-                      >
-                        <Icon icon={step.icon} />
-                      </RenderAvatar>
-                      <div>
-                        <Typography className='step-title'>{step.title}</Typography>
-                        <Typography className='step-subtitle'>{step.subtitle}</Typography>
-                      </div>
-                    </div>
-                  </StepLabel>
-                </Step>
-              )
-            })}
-          </Stepper>
-        </StepperWrapper>
-      </CardContent>
-      <Divider sx={{ m: '0 !important' }} />
-      <CardContent>{renderContent()}</CardContent>
-    </Card>
-  )
-}
-
-export default StepperCustomHorizontal
