+++ src/SharedComponents/PickersCustomInput.tsx
@@ -1,0 +1,21 @@
+// ** React Imports
+import { forwardRef } from 'react'
+
+// ** MUI Imports
+import TextField from '@mui/material/TextField'
+
+interface PickerProps {
+  label?: string
+  readOnly?: boolean
+}
+
+const PickersComponent = forwardRef(({ ...props }: PickerProps, ref) => {
+  // ** Props
+  const { label, readOnly } = props
+
+  return (
+    <TextField inputRef={ref} {...props} label={label || ''} {...(readOnly && { inputProps: { readOnly: true } })} />
+  )
+})
+
+export default PickersComponent
